+++ client/src/App.tsx
@@ -1,52 +1,20 @@
 import React from 'react';
-import Header from './components/Header';
-import Rating from './components/Rating';
-import Footer from './components/Footer';
+import { Route, Redirect, Switch } from 'react-router-dom';
+import Auth from './pages/Auth/Auth';
+import Country from './pages/Country/Country';
+import Main from './pages/Main/Main';
 
 const App = () => {
-  // HARDCODE
-  const RenderRating: React.FC = () => {
-    const attractionRating = [
-      {
-        stars: 1,
-        numReviews: 3,
-      },
-      {
-        stars: 1.5,
-        numReviews: 12,
-      },
-      {
-        stars: 3,
-        numReviews: 6,
-      },
-      {
-        stars: 2,
-        numReviews: 9,
-      },
-      {
-        stars: 5,
-        numReviews: 54,
-      }
-    ];
-
-    return (
-      <div>
-        {attractionRating.map(el => (
-          <Rating value={el.stars} text={`${el.numReviews} reviews`} />
-        ))}
-      </div>
-    )
-  }
-
   return (
     <>
-      <Header />
-      <main>
-        <RenderRating />
-      </main>
-      <Footer />
+      <Switch>
+        <Route path={'/main'} component={Main} exact />
+        <Route path={'/country'} component={Country} exact />
+        <Route path={'/auth'} component={Auth} exact />
+        <Redirect to={'/main'} />
+      </Switch>
     </>
   );
-}
+};
 
 export default App;
